PlutoNotebook now imported by default in notebooks
