+++ src/Pluto.jl
@@ -51,7 +51,7 @@ include("./webserver/Static.jl")
 include("./webserver/Dynamic.jl")
 include("./webserver/REPLTools.jl")
 include("./webserver/WebServer.jl")
-include("./webserver/REST.jl")
+include("./rest/REST.jl")
 
 if get(ENV, "JULIA_PLUTO_SHOW_BANNER", "1") !== "0"
 @info """\n
