+++ src/runner/PlutoRunner.jl
@@ -21,6 +21,7 @@ import UUIDs: UUID
 import Logging
 
 export @bind
+export PlutoNotebook
 
 MimedOutput = Tuple{Union{String,Vector{UInt8},Dict{Symbol,Any}},MIME}
 ObjectID = typeof(objectid("hello computer"))
@@ -1194,6 +1195,13 @@ end"""
 
 
 
+
+
+###
+# REST (move elsewhere?)
+###
+
+include("../rest/RESTClient.jl")
